pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf_yaml_bookmark"
version = "0.0.1"
authors = [
  { name = "Yuta Yamamoto", email = "jichael.y.yuta1995@gmail.com" },
]
description = "Adds bookmarks to PDF documents."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pyyaml",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/yutayamamoto/pdfoutline"
Issues = "https://github.com/yutayamamoto/pdfoutline/issues"
